Create interaction page broken down into components
