+++ frontend/src/components/CreateInteraction/GradeSlider.tsx
@@ -1,5 +1,5 @@
-import {GradeComponent} from "../../api/interaction.ts";
-import {useState} from "react";
+import { GradeComponent } from "../../api/interaction.ts";
+import { useState } from "react";
 
 const GradeSlider = ({
                          component,
@@ -16,7 +16,6 @@ const GradeSlider = ({
     const [inputValue, setInputValue] = useState<string>(value?.toString() ?? component.minGrade.toString());
     const fullPath = path ? `${path}.${component.name}` : component.name;
 
-    // Update the parent component when slider changes
     const handleSliderChange = (e: React.ChangeEvent<HTMLInputElement>) => {
         const newValue = parseFloat(e.target.value);
         setSliderValue(newValue);
@@ -24,15 +23,12 @@ const GradeSlider = ({
         onChange(fullPath, newValue);
     };
 
-    // Validate and update when text input changes
     const handleInputChange = (e: React.ChangeEvent<HTMLInputElement>) => {
         const newValue = e.target.value;
         setInputValue(newValue);
 
-        // Validate if it's a valid number and within range
         const numValue = parseFloat(newValue);
         if (!isNaN(numValue) && numValue >= component.minGrade && numValue <= component.maxGrade) {
-            // Check if the value is valid according to step
             const stepsFromMin = (numValue - component.minGrade) / component.stepAmount;
             const isValidStep = Math.abs(Math.round(stepsFromMin) - stepsFromMin) < 0.0001;
 
@@ -43,18 +39,19 @@ const GradeSlider = ({
         }
     };
 
-    // Validate on blur to correct any invalid input
     const handleBlur = () => {
-        // Round to nearest valid step
         const stepsFromMin = (sliderValue - component.minGrade) / component.stepAmount;
         const roundedSteps = Math.round(stepsFromMin);
         const validValue = component.minGrade + (roundedSteps * component.stepAmount);
-        const fixedValue = Number(validValue.toFixed(2)); // Fix floating point precision issues
+        const fixedValue = Number(validValue.toFixed(2));
 
         setSliderValue(fixedValue);
         setInputValue(fixedValue.toString());
         onChange(fullPath, fixedValue);
     };
+
+    // Calculate percentage for dynamic styling
+    const percentage = ((sliderValue - component.minGrade) / (component.maxGrade - component.minGrade)) * 100;
 
     return (
         <div className="mb-4 bg-white p-4 rounded-md shadow-sm border border-gray-200">
@@ -85,18 +82,25 @@ const GradeSlider = ({
 
             <div className="flex items-center">
                 <span className="text-xs text-gray-500 mr-2">{component.minGrade}</span>
-                <input
-                    type="range"
-                    min={component.minGrade}
-                    max={component.maxGrade}
-                    step={component.stepAmount}
-                    value={sliderValue}
-                    onChange={handleSliderChange}
-                    className="flex-grow h-2 appearance-none rounded-lg bg-gray-200 accent-primary-600 cursor-pointer"
-                />
+                <div className="relative w-full">
+                    <input
+                        type="range"
+                        min={component.minGrade}
+                        max={component.maxGrade}
+                        step={component.stepAmount}
+                        value={sliderValue}
+                        onChange={handleSliderChange}
+                        className="w-full h-2 appearance-none bg-transparent cursor-pointer"
+                        style={{
+                            background: `linear-gradient(to right, #7a24ec 0%, #7a24ec ${percentage}%, #e5e7eb ${percentage}%, #e5e7eb 100%)`,
+                            borderRadius: '9999px'
+                        }}
+                    />
+                </div>
                 <span className="text-xs text-gray-500 ml-2">{component.maxGrade}</span>
             </div>
         </div>
     );
 };
+
 export default GradeSlider;
