+++ frontend/src/components/CreateInteraction/InteractionSuccessMessage.tsx
@@ -1,0 +1,7 @@
+const InteractionSuccessMessage = () => (
+    <div className="fixed top-4 right-4 bg-green-100 border border-green-400 text-green-700 px-4 py-3 rounded z-50 shadow-md">
+        Your interaction has been submitted successfully!
+    </div>
+);
+
+export default InteractionSuccessMessage;
